Have `detect` accept `&[Joint]` rather than `Skeleton` for flexibility
Only accepting `Skeleton` makes it tricky to do detection on a thread
other than the one on which the callbacks are run. This PR allows for
doing detection on both the skeletons produced via `SkeletonData` as
well as user owned skeleton types, as long as a slice of `Joint`s is
accessible.